+++ src/resources/auth/auth.router.ts
@@ -1,10 +1,10 @@
 import { Router } from "express";
-import { createSessionToken } from "utils/auth";
 import { StatusCodes } from "http-status-codes";
 import { handleRequest } from "../../utils/handleRequest";
 import User from "../../entities/user.entity";
 
-import * as loginService from "./login.service";
+import * as authService from "./auth.service";
+import { createSessionToken } from "./auth.service";
 
 const router = Router();
 
@@ -12,7 +12,7 @@ router
   .route('/')
   .post(handleRequest(async (req, res) => {
     const { login, password } = req.body;
-    const auth = await loginService.authorizeUser(login, password);
+    const auth = await authService.authenticate(login, password);
 
     if (typeof auth === 'string') {
       res
